requirements.txt
Scrapy==1.4.0
scrapy-splash==0.7.2
selenium-3.4.3
